requirements.txt
ultralytics
python-dotenv
numpy
opencv-python
